script: Avoid panic when scrolling area of window is larger than viewport (#39367)
Sometimes, the computed scrolling area of window is larger than
viewport. This causes panics in `Window.scroll` with `f32::clamp(0.0,
some negative number)`.

Eventually, we should find out why "computed scrolling area of window is
larger than viewport". But let's avoid the panics first.

Testing: This avoids panic, so definitely not covered by existing tests.
But it would be hard to write a automated test for this in headless
mode.
Fixes: #39346

Signed-off-by: Euclid Ye <yezhizhenjiakang@gmail.com>