script: Implement methods of IDBCursor
Implement the `advance`, `continue`, `continuePrimaryKey`, `update` and
`delete` methods of the `IDBCursor` interface.

All the steps are implemented except those calling
`IDBRequest::execute_async`. According to the specification, we should
execute a request with the *object store*, *index*, or *cursor* as its
source. Currently, `IDBRequest::execute_async` only accepts using an
object store as source. In this patch,

- For the `advance`, `continue` and `continuePrimaryKey` methods, the
  case where an index is used as source has not yet implemented.
- For the `update` and `delete`, we temporarily use the cursor's
  effective object store, as a temporary workaround. This is expected to
  be corrected in the future.

Signed-off-by: Kingsley Yung <kingsley@kkoyung.dev>