constellation: Store pending viewport changes (#39447)
It is possible for viewport changes to be received before the
constellation adds a browsing context to its managed set when
navigating. When that happens, the viewport change is ignored, and eg.
the hidpi will keep the default 1.0 value.
On later page navigation, this value is not updated and leads to
incorrect display zooming. The initial page load is fine in the current
embedders because they update the compositor and webview renderer even
if the constellation is not fully ready.

Testing: Testing is difficult because it's device dependent, but I
reproduce 100% of the time running winit_minimal on a Pinephone Pro.

Signed-off-by: webbeef <me@webbeef.org>