servoshell: Scale WebDriver mouse button coordinates by HiDPI scale factor
Mouse button events are sent in CSSPixel coordinates, but these need to
so they need to converted into device coordinates when creating an input
event for them. This isn't an issue for automated test because they
always use a scale factor of 1. This is a problem when running
interactively and not in headless mode.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>