servoshell: Do not override HiDPI ratio when taking screenshots
This override is ostensibly so that reftests always run with a device
pixel ratio of 1.0. Reftests are always run in headless mode which uses
a device pixel ratio of 1.0 regardless (unless you override it).
Removing this particular override allows taking screenshots in the
system DPI when taking them in headed mode, which is actually a bit
useful.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>