script: Migrate "encrypt"/"decrypt" operation to use new normalization
Refactoring of the algorithm normalization in #_____ introduces a new
algorithm normalization procedure to replace the existing one, and we
continue the migration here.

This patch migrates the "encrypt" and "decrypt" operation from using
existing `normalize_algorithm_for_encrypt_or_decrypt` function to using
the new `normalize_algorithm` function.

We are expected to move

- `SubtleCrypto::encrypt_aes_cbc`,
- `SubtleCrypto::decrypt_aes_cbc`,
- `SubtleCrypto::encrypt_decrypt_aes_ctr`,
- `SubtleCrypto::encrypt_aes_gcm` and
- `SubtleCrypto::decrypt_aes_gcm`

to the sub-module `aes_operation`. However, they are only copied to the
sub-module without removing them. This is because the `wrapKey` and
`unwrapKey` operation (not yet migrated) depend on them. When the two
operations are also migrated, we can remove them.

Testing: Refactoring. Existing WPT tests are enough.
Fixes: Part of #_____

Signed-off-by: Kingsley Yung <kingsley@kkoyung.dev>