script: Return cached object from `CryptoKey.usages` getter
Create and store a cached object of the `[[usages]]` internal slot for
`CryptoKey` when the key is created or when its `[[usages]]` internal
slot is updated. The getter can then return the cached object as
specified in https://w3c.github.io/webcrypto/#dom-cryptokey-usages,
instead of creating a new object on each call.

Signed-off-by: Kingsley Yung <kingsley@kkoyung.dev>