layout: Ensure that the `Epoch` stored in layout is accurate
The first epoch is 0 as that is the one used in the initial transaction,
but the code was setting the first `Epoch` to `Epoch(1)`. This means that
when layout advanced the epoch, the `Epoch` of the first produced
display list was `Epoch(2)`.

This change makes the value reflected in `current_epoch` actually match
the index of the display list produced. In addition, we always store
this epoch in `PipelineDetails` in the renderer. This will be important
when adding support for adding the `WebView::take_screenshot` API.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>