layout: Clone static position rectangles when caching in IndependentFormattingContext
Hoisted absolutes that are cached in IndependentFormattingContexts may
have their HoistedSharedFragment resused between layouts. This means
that the original value of the adjusted static positioning rectangles
need to be preserved in cache. This change makes it so that these values
are cloned when being retrieved from the cache. Further adjustments up
the tree will now no longer affected the version in teh cache.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>
Co-authored-by: Oriol Brufau <obrufau@igalia.com>