Update expectations for `--product servodriver` (#39603)
this patch updates the expectations for tests that use testdriver.js,
which currently only work with `--product servodriver` (#34683), running
the tests with #39587.

expectations for all other tests, which don’t use testdriver.js, are
unaffected.

Testing: this patch updates test expectations, though the tests are not
yet run in CI

Signed-off-by: Delan Azabani <dazabani@igalia.com>
Co-authored-by: Martin Robinson <mrobinson@igalia.com>