Add client to request
This aligns the request object more with the specification,
since the spec now has a `traversable_for_user_prompts` and
a separate field for the client. Before, they were present
in the same enum.

In doing so, new structs are added that are all required in
the new spec. It also adds a minimal client to start with
the concept of preloaded resources.

Since the spec moved things around a bit, it now has a
dedicated method to populate the client from the request.

Part of #35035

Signed-off-by: Tim van der Lippe <tvanderlippe@gmail.com>