wpt: Capture test output when running WPT in WebDriver mode (#39656)
We were capturing test output (such as stack traces) by using the "test"
key in the data provided to the `process_output` log handler. This
doesn't work for WebDriver mode because the test runner runs in another
process entirely and it may reuse Servo executable instances for tests.

This change maps browser process output to tests using the "browser_pid"
key that is often provided while in WebDriver mode.

Testing: This modifies the test harness, so the tests are running tests.
Fixes: #39176

Signed-off-by: Martin Robinson <mrobinson@igalia.com>