Experiment: put DOM object implementation in separate crate
This is an experiment if it is feasible to put DOM object
implementations into a separate crate. With some wrangling,
it was possible to compile in a separate crate, when we
comment out the `reflect_dom_object` since that refers to
the `GlobalScope`.

Then I tried to modify the WebIDL setup to take into account
the different scripts. It can now compile the full WebIDL,
but that required some moving things around. As a result of
that, while it parses the WebIDL, it produces Rust code
that doesn't compile yet.

Putting this out there to see what folks think of this approach.
Ultimately the goal is to produce a leaner build graph so
that we can speed up build times.

Signed-off-by: Tim van der Lippe <tvanderlippe@gmail.com>