compositor: Fix a performance regression from #39583 (#39684)
When processing new WebRender frames after handling messages in the
compositor, only do that when we actually receive
`NewWebRenderFrameReady` messages. This isn't an issue unless there are
active animations. When there are, every spin of the event loop was
triggering a repaint.

Testing: This should fix the performance regression on the Speedometer
Bencher
benchmarks.
Fixes: #39641

Signed-off-by: Martin Robinson <mrobinson@igalia.com>
Co-authored-by: Delan Azabani <dazabani@igalia.com>