layout: Unify line height resolution (#39688)
Both `flow/inline/line.rs` and `flow/inline/mod.rs` had a function to
resolve the `line-height` value. They were pretty much the same, except
that the former wasn't handling single-line text inputs. But I don't
think it matters, since the text inside such input can't be styled with
`vertical-align: top` or `bottom`.

Thus this patch unifies the logic. And to simplify the callers, now this
function accepts a `&InlineContainerStateFlags` instead of a bool for
identifying single-line text inputs.

Testing: There shouldn't be any behavior change.
This is part of #39664

Signed-off-by: Oriol Brufau <obrufau@igalia.com>