Add `test-wait` to `/_mozilla/shadow-dom/move-element-with-ua-shadow-tree-crash.html` (#39476)
Without the `test-wait` class the testharness won't wait until the
actual test executes.

I'm not sure if this is the reason for
https://github.com/servo/servo/issues/39473 (I don't understand how that
could cause the test to time out), but it's definitely *a* bug. The test
is not intermittent locally. Let's do a couple try runs and then see
where we go from there.

Signed-off-by: Simon Wülker <simon.wuelker@arcor.de>