webdriver: Send events to the embedder as `InputEvent` with a response channel
This change simplifies the WebDriver <-> Embedder API by exposing a
single message for dispatch input event actions. This message now
includes a `Sender` which is used to inform the WebDriver server that
the embedder has finished processing the message in the DOM. This
replaces the system of ids and the side channel that was used for this
purpose before.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>