Add policy container to all requests
In preparation of adding a client to a request, we first
need to add a policy container to all requests. Some
where missing and relying on fallbacks (that need to
be removed).

For worklets this adds the appropriate policy container
from the containing global, but since that's still behind
a pref, no tests are affected.

For font contexts I tried to pass through the relevant
information, but it is a tad too much. Once fontcontext
knows more information about its containing global (as
the TODO also points to referrer), we should clean this up.

Part of #35035

Signed-off-by: Tim van der Lippe <tvanderlippe@gmail.com>