xpath: Avoid allocating hashmap when comparing NodeSet and String (#39774)
A NodeSet and a String are equal iff the text content of any node in the
set is equal to the string. Instead of creating a hashmap of text
contents and then calling `contains` on that, just iterate over the
nodeset and see if the string is in there.

Testing: This should not change behaviour and is thus covered by
existing tests
Part of #34527

Signed-off-by: Simon Wülker <simon.wuelker@arcor.de>