servoshell: Reimplement "onClick" action on Android (#39771)
The Android interface needs some way to communciate click events to
Servo, so the JNI layer needs to expose a click event. Even though there
is no longer a "click" action in the API, we can implement this behavior
via a mouse button down and then mouse button up event, which should be
a more complete implementation in the DOM.

This fixes a regression from #39705.

Testing: There currently aren't tests for Android interaction via the
Java layer
(which is one reason why this regression happened sadly).
Fixes #39742.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>