libservo: Notify the embedder when an input event can't be seent to a Pipeline
The `WebViewDelegate::notify_input_event_handled` method is called for the
embedder when an input event's processing has finished within Servo.
This allows the embedder to chain the event up to parent widgets or do
some kind of post-processing such as the implementation of overridable
keybindings.

This change makes it so that this method is called even when an input
event cannot successfully be sent to a pipeline. This could happen when
an event is handled during navigation or perhaps during shutdown. In any
case, Servo should *always* tell the embedder when the event is done
processing so that they don't get eaten by the Servo widget.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>