script: Access internal slot of CryptoKey from HMAC Operations
After the PR #_____ being merged, we can access the `[[algorithm]]`
internal slot of a `CryptoKey` directly in Rust, without doing a JS
object conversion.

This patch leverages this feature to simplify the steps that access the
`[[algorithm]]` internal slot of `CryptoKey`, in HMAC operations.

Testing: Refactoring. Existing WPT tests suffice.

Signed-off-by: Kingsley Yung <kingsley@kkoyung.dev>