constellation: prevent panic when constellation receives message from closed pipeline (#39745)
As part of https://github.com/servo/servo/issues/31973, I noticed that
the intermittent crash of
`/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_popups_escaping-3.html`
appears due to the closing of the of the source webview of the message,
which prior to these changes would result in an early return and drop
the reply sender, resulting in a panic.

This change makes the check for whether the webview is still open
specific to certain messages that need the webview and
`GetBrowsingContextInfo` is not one of them.

As to exactly what allows a webview to be closed and still send a
`GetBrowsingContextInfo` message, it appears that the iframe and the
pop-up opened by the iframe share the same webview, and when the pop-up
calls `close`, that closes the webview but the iframe is still running
and then receives a post message from a webview that was already
closed(intermittently). See https://github.com/servo/servo/issues/39748

Testing: WPT
Fixes: https://github.com/servo/servo/issues/22647

---------

Signed-off-by: gterzian <2792687+gterzian@users.noreply.github.com>